mnesia: copy table from a specified node
When a new node is added to the mnesia cluster, it selects a random node
to copy the table from if no 'master' node is defined (set master node
is not mandatory).

by setting arg: 'copy_from_node', operator is now free to choice
which node to copy from.

default is 'undefined', means disable.

note, if the specified node is not in the candidates list, it will fallback to
default behavior.